pyproject.toml
[project]
name = "hello-pulsar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pulsar-client>=3.6.1",
]
